[vulkan] Split the main page into three pages.
The main page was starting to be a bit loaded. I had to decide how to
present the information now that Vulkan is here.

The design I chose is that the main page display the commits and the
statistics, and the OpenGL and Vulkan pages show the details of each
driver.

With this commit comes some kind of Controller design: the web pages
(i.e. index, opengl, vulkan, drivers and about) simply calls specific
Controllers. Later, there could be a hub to instantiate the appropriate
Controller depending on the wanted page.

Controller allows to structure and factorize a bit the HTML rendering.
All the Controllers must extend the BaseController abstract class and
implement computeRendering() and writeHtmlPage(). These two methods are
called at specific moments in the page rendering.